Refactor Flask Routes and Updated NotificationFeedPaje to authenticate user
