+++ frontend-react-js/src/pages/NotificationsFeedPage.js
@@ -44,7 +44,7 @@ export default function NotificationsFeedPage() {
     dataFetchedRef.current = true;
 
     loadData();
-    checkAuth();
+    checkAuth(setUser);
   }, []);
 
   return (
